cleanup Nfc fakes/fixture files. Introduced forward declaration of classes. Updated Nfc to only ask for Folder's intialization, not validating it.
