+++ test/desktop/mocks/unittest_NfcTag_mocks.h
@@ -4,7 +4,31 @@
 #include <gmock/gmock.h>
 #include <Arduino_types.h>
 #include <MFRC522_interface.h>
-#include "unittest_NfcTag_fakes.h"
+#include <NfcTag.h>
+
+// FAKES
+// Fake buffer data for NFC tag read
+static const byte fakeBufferData[16]{
+    (byte)(NfcTag::cui32MagicCookie >> 24),          // 0
+    (byte)((NfcTag::cui32MagicCookie >> 16) & 0xFF), // 1
+    (byte)((NfcTag::cui32MagicCookie >> 8) & 0xFF),  // 2
+    (byte)(NfcTag::cui32MagicCookie & 0xFF),         // 3
+    (byte)1,                                         // 4 FolderId
+    (byte)Folder::LULLABYE,                          // 5 ePlayMode
+    (byte)5,                                         // 6 TrackCount
+    0, 0, 0, 0, 0, 0, 0, 0, 0};
+
+class Fake_MFRC522 : public MFRC522_interface
+{
+public:
+    virtual ~Fake_MFRC522() {}; // MUST BE DEFINED; ELSE VTABLE INCLUDE ERRORS
+    void initReader() override;
+    bool isCardPresent() override;
+    bool isNewCardPresent() override;
+    bool writeCard(byte blockAddr, byte *dataToWrite) override;
+    bool readCard(byte blockAddr, byte *readResult) override;
+};
+
 
 // MOCKS
 class Mock_MFRC522 : public MFRC522_interface
@@ -29,4 +53,21 @@ public:
 private:
     Fake_MFRC522 m_FakeRead {};
 };
+
+
+
+// MATCHERS
+MATCHER_P2(arrayByteCompare, bytes, size, "Compares array bites and throws errors for each byte that does not match.")
+{
+    for (int i = 0; i < size; ++i)
+    {
+        if (arg[i] != bytes[i])
+        {
+            EXPECT_EQ("", "Error: " + std::to_string(i) + ". element not matching:");
+            EXPECT_EQ(arg[i], bytes[i]);
+        }
+    }
+    return true;
+}
+
 #endif // UNITTEST_NFCTAG_MOCKS_H
